pyproject.toml
[tool.poetry]
name = "infrallm"
version = "0.1.0"
description = "AI-powered self-service infrastructure provisioning for platform engineers"
authors = ["Platform Engineering Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
anthropic = "^0.39.0"
PyGithub = "^2.1.1"
click = "^8.1.7"
pyyaml = "^6.0.1"
rich = "^13.7.0"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
ruff = "^0.1.9"

[tool.poetry.scripts]
infrallm = "src.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
